requirements.txt
requests
dnspython
pygeoip
